#27 - 컨트롤러 코드와 주석 분리를 위한 interface 도입
